% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emoji_rain_dependency}
\alias{emoji_rain_dependency}
\title{Emoji Rain Assets Dependency}
\usage{
emoji_rain_dependency()
}
\value{
An htmlDependency object
}
\description{
Emoji Rain Assets Dependency
}
\keyword{internal}
