% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{Null Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left operand}

\item{y}{Right operand (default value)}
}
\value{
x if x is not NULL, otherwise y
}
\description{
Returns the left operand if it is not NULL, otherwise returns the right operand.
}
\keyword{internal}
